requirements.txt
flask
python-dotenv
tinydb
pytz
APScheduler
pycryptodome